+++ include/types.h
@@ -5,13 +5,15 @@
 #ifndef __TYPES_H__
 #define __TYPES_H__
 
+#include <stdint.h>
+
 // Machine types
 typedef unsigned char		Byte;		/* 8 bits */
 typedef unsigned short		SWord;		/* 16 bits */
-typedef unsigned int		DWord;		/* 32 bits */
-typedef unsigned int		dword;		/* 32 bits */
-typedef unsigned int		Word;		/* 32 bits */
-typedef unsigned int		ADDRESS;	/* 32-bit unsigned */
+typedef unsigned int			DWord;		/* 32 bits */
+typedef unsigned int			dword;		/* 32 bits */
+typedef unsigned int			Word;		/* 32 bits */
+typedef uintptr_t					ADDRESS;	/* pointer. size depends on platform */
 
 
 #define STD_SIZE	32					// Standard size
